% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_inputs.R
\name{assert_vector}
\alias{assert_vector}
\title{Assert that an input is a vector with desired properties}
\usage{
assert_vector(x, class, len = NULL, null_ok = FALSE)
}
\arguments{
\item{x}{A vector. The input to be analysed.}

\item{class}{A string. The class that this input should inherit from.}

\item{len}{An integer. The length of the vector. If \code{NULL}, length is not
checked.}

\item{null_ok}{A logical. Whether the input could also be \code{NULL}.}
}
\value{
\code{TRUE} if the check is successful. Throws an error describing the
issue with the input otherwise.
}
\description{
Assert that an input is a vector with desired properties
}
\keyword{internal}
