% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_subset.R
\name{check_invalid_value}
\alias{check_invalid_value}
\title{Check for invalid values}
\usage{
check_invalid_value(x, value)
}
\arguments{
\item{x}{A GTFS object.}

\item{value}{Either a numeric or a character vector.}
}
\value{
\code{NULL}, invisibly. Called for side-effects.
}
\description{
Checks if all elements from \code{value} correspond to an element from
\code{x} and throws a warning if any invalid value is found.
}
\keyword{internal}
