% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gtfs_standards.R
\name{get_gtfs_standards}
\alias{get_gtfs_standards}
\title{Generate GTFS standards (deprecated)}
\usage{
get_gtfs_standards()
}
\value{
A named list, in which each element represents the R equivalent of
each GTFS table standard (based on the specifications of 2022-05-09).
}
\description{
This function is \strong{deprecated} and no longer used in \code{\link[=import_gtfs]{import_gtfs()}} or \code{\link[=export_gtfs]{export_gtfs()}}.
The dataset \link{gtfs_reference} now contains the standard specifications.
}
\seealso{
\link{gtfs_reference}
}
