% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfsio.R
\docType{package}
\name{gtfsio-package}
\alias{gtfsio}
\alias{gtfsio-package}
\title{gtfsio: Read and Write General Transit Feed Specification (GTFS) Data}
\description{
Tools for the development of GTFS-related packages. Establishes a standard
for representing GTFS feeds using R data types. Provides fast and flexible
functions to read and write GTFS feeds while sticking to this standard.
Defines a basic \code{gtfs} class which is meant to be extended by packages
that depend on it. And offers utility functions that support checking the
structure of GTFS objects.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r-transit.github.io/gtfsio/}
  \item \url{https://github.com/r-transit/gtfsio}
  \item Report bugs at \url{https://github.com/r-transit/gtfsio/issues}
}

}
\author{
\strong{Maintainer}: Flavio Poletti \email{flavio.poletti@hotmail.ch}

Authors:
\itemize{
  \item Daniel Herszenhut \email{dhersz@gmail.com} (\href{https://orcid.org/0000-0001-8066-1105}{ORCID})
  \item Mark Padgham \email{mark.padgham@email.com}
}

Other contributors:
\itemize{
  \item Rafael H. M. Pereira \email{rafa.pereira.br@gmail.com} (\href{https://orcid.org/0000-0003-2125-7465}{ORCID}) [reviewer]
  \item Tom Buckley \email{tom@tbuckl.com} [reviewer]
  \item Ipea - Institute for Applied Economic Research [copyright holder, funder]
}

}
\keyword{internal}
