% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfsio.R
\docType{package}
\name{gtfsio}
\alias{gtfsio}
\alias{_PACKAGE}
\alias{gtfsio-package}
\title{gtfsio: Read and Write General Transit Feed Specification (GTFS) Data}
\description{
Provides a 'gtfs' class to represent General
    Transit Feed Specification (GTFS) files, and fast and flexible tools
    to read and write such files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-transit/gtfsio}
  \item Report bugs at \url{https://github.com/r-transit/gtfsio/issues}
}

}
\author{
\strong{Maintainer}: Daniel Herszenhut \email{dhersz@gmail.com} (\href{https://orcid.org/0000-0001-8066-1105}{ORCID})

Other contributors:
\itemize{
  \item Flavio Poletti \email{flavio.poletti@hotmail.ch} [contributor]
  \item Mark Padgham \email{mark.padgham@email.com} [contributor]
  \item Rafael H. M. Pereira \email{rafa.pereira.br@gmail.com} (\href{https://orcid.org/0000-0003-2125-7465}{ORCID}) [contributor]
  \item Tom Buckley \email{tom@tbuckl.com} [contributor]
}

}
