% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_gtfs.R
\name{read_geojson}
\alias{read_geojson}
\title{Read geojson file}
\usage{
read_geojson(file.geojson)
}
\arguments{
\item{file.geojson}{geojson file}
}
\value{
json list
}
\description{
Read geojson file
}
\keyword{internal}
