% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_subset.R
\name{subset.gtfs}
\alias{subset.gtfs}
\alias{$.gtfs}
\alias{[.gtfs}
\alias{[[.gtfs}
\title{Subset a GTFS object}
\usage{
\method{$}{gtfs}(x, name)

\method{[}{gtfs}(x, value)

\method{[[}{gtfs}(x, value)
}
\arguments{
\item{x}{A GTFS object.}

\item{name}{A string. The name of the element to be returned.}

\item{value}{Either a numeric or a character vector. Designates the elements
to be returned.}
}
\value{
A GTFS object.
}
\description{
Subsetting a GTFS object using \code{[}, \code{[[} and \code{$} is very
similar to subsetting lists, but a bit more verbose, as warnings are thrown
when attempting to subset a missing element. When using \code{[} the
resulting object inherits classes (and subclasses) from the original object.
}
\examples{
gtfs_path <- system.file("extdata/ggl_gtfs.zip", package = "gtfsio")

gtfs <- import_gtfs(gtfs_path)
names(gtfs)
class(gtfs)

# '[' behaviour
small_gtfs <- gtfs[1:5]
names(gtfs)
class(small_gtfs)

bad_gtfs <- gtfs["ola"] # results in warning

# '$' behaviour
stop_times <- gtfs$stop_times
names(stop_times)

bad_element <- gtfs$ola # results in warning

# '[[' behaviour
shapes <- gtfs$shapes
names(shapes)

bad_element <- gtfs[["ola"]] # results in warning
}
\concept{subset}
