% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs_methods.R
\name{summary.gtfs}
\alias{summary.gtfs}
\title{Print summary of a GTFS object}
\usage{
\method{summary}{gtfs}(object, ...)
}
\arguments{
\item{object}{A GTFS object.}

\item{...}{Ignored.}
}
\value{
Vector with the number of entries for each gtfs table.
}
\description{
Print summary of a GTFS object
}
\examples{
gtfs_path <- system.file("extdata/ggl_gtfs.zip", package = "gtfsio")
gtfs <- import_gtfs(gtfs_path)

summary(gtfs)

}
