% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gtfs_standards.R
\name{translate_types}
\alias{translate_types}
\title{Translate GTFS specification types to R equivalent types}
\usage{
translate_types(text_file, r_equivalents)
}
\arguments{
\item{text_file}{A named \code{list} containing a GTFS text file
specification, as described in the body of
\code{\link{get_gtfs_standards}}.}

\item{r_equivalents}{A named \code{character vector}, in which each name is
the GTFS specification type and the content its R equivalent.}
}
\value{
A named \code{list} holding a GTFS text file specification, but with
  R data types instead of GTFS spec data types.
}
\description{
Translate GTFS specification types to R equivalent types
}
\keyword{internal}
