% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert_inputs.R
\name{assert_vector}
\alias{assert_vector}
\alias{assert_list}
\title{Assert that an input is a vector/list with desired properties}
\usage{
assert_vector(
  x,
  class,
  len = NULL,
  null_ok = FALSE,
  var_name = NULL,
  n_call = -1L
)

assert_list(x, len = NULL, null_ok = FALSE)
}
\arguments{
\item{x}{A vector. The input to be analysed.}

\item{class}{A string. The class that this input should inherit from.}

\item{len}{An integer. The length of the vector. If \code{NULL}, length is not
checked.}

\item{null_ok}{A logical. Whether the input could also be \code{NULL}.}

\item{var_name}{A string. The name of the variable in the informative
message. If \code{NULL}, the name is guessed from the function call.}

\item{n_call}{A negative integer. The number of frames to go back to find the
call to associate the error generated by the function with. Defaults to the
parent frame (-1). This argument is only relevant to other assertion
functions that build on top of this one.}
}
\value{
\code{TRUE} if the check is successful. Throws an error describing the
issue with the input otherwise.
}
\description{
Assert that an input is a vector/list with desired properties
}
\concept{input assertion}
\keyword{internal}
