% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gtfs_standards.R
\name{get_gtfs_standards}
\alias{get_gtfs_standards}
\title{Generate GTFS standards}
\usage{
get_gtfs_standards()
}
\value{
A named list, in which each element represents the R equivalent of
each GTFS table standard.
}
\description{
Generates a list specifying the standards to be used when reading and writing
GTFS feeds with R. Each list element (also a list) represents a distinct GTFS
table, and describes:
\itemize{
\item the fields that compose the table, including which R data type is best
suited to represent it, whether the field is required, optional or
conditionally required.
\item whether the table is required, optional or conditionally required (as an
attribute)
}
}
\section{Details}{

GTFS standards were derived from \href{https://gtfs.org/schedule/reference/}{GTFS Schedule Reference}. The R data types chosen to
represent each GTFS data type are described below:
\itemize{
\item{Unique ID = \code{character}}
\item{Text = \code{character}}
\item{URL = \code{character}}
\item{Timezone = \code{character}}
\item{Language code = \code{character}}
\item{Phone number = \code{character}}
\item{Email = \code{character}}
\item{ID = \code{character}}
\item{Color = \code{character}}
\item{Time = \code{character}}
\item{Currency code = \code{character}}
\item{String = \code{character}}
\item{Enum = \code{character}}
\item{Text or URL or Email or Phone number = \code{character}}
\item{Array = \code{geojson_array}}
\item{Object = \code{geojson_object}}
\item{Enum = \code{integer}}
\item{Non-negative integer = \code{integer}}
\item{Date = \code{integer}}
\item{Non-zero integer = \code{integer}}
\item{Positive integer = \code{integer}}
\item{Non-null integer = \code{integer}}
\item{Integer = \code{integer}}
\item{Latitude = \code{numeric}}
\item{Longitude = \code{numeric}}
\item{Non-negative float = \code{numeric}}
\item{Currency amount = \code{numeric}}
\item{Float = \code{numeric}}
\item{Positive float = \code{numeric}}
}
}

\examples{
gtfs_standards <- get_gtfs_standards()

}
