% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gtfs_reference}
\alias{gtfs_reference}
\title{GTFS reference}
\format{
A list with data for every GTFS file. Each named list element (also a list) has
specifications for one GTFS file in the following structure:
\itemize{
\item{\code{File_Name}: file name including file extension (txt or geojson)}
\item{\code{File_Presence}: Presence condition applied to the file}
\item{\code{file}: file name without file extension}
\item{\code{file_ext}: file extension}
\item{\code{fields}: data.frame with parsed field specification (columns:
\code{Field_Name}, \code{Type}, \code{Presence}, \code{Description}, \code{gtfsio_type})}
\item{\code{primary_key}: primary key as vector}
\item{\code{field_types}: named vector on how GTFS types (values) should be read in gtfsio
(names). Values are the same as in \code{fields}.}
}
}
\source{
\url{https://github.com/google/transit/blob/master/gtfs/spec/en/reference.md}
}
\usage{
gtfs_reference
}
\description{
The data from the official GTFS specification document parsed to a list.
}
\details{
GTFS Types are converted to R types in gtfsio according to the following list:
\itemize{
\item{Unique ID = \code{character}}
\item{Text = \code{character}}
\item{URL = \code{character}}
\item{Timezone = \code{character}}
\item{Language code = \code{character}}
\item{Phone number = \code{character}}
\item{Email = \code{character}}
\item{Time = \code{character}}
\item{Currency code = \code{character}}
\item{ID = \code{character}}
\item{String = \code{character}}
\item{Color = \code{character}}
\item{Enum = \code{character}}
\item{Text or URL or Email or Phone number = \code{character}}
\item{Array = \code{geojson_array}}
\item{Object = \code{geojson_object}}
\item{Enum = \code{integer}}
\item{Integer = \code{integer}}
\item{Date = \code{integer}}
\item{Non-negative integer = \code{integer}}
\item{Non-zero integer = \code{integer}}
\item{Positive integer = \code{integer}}
\item{Non-null integer = \code{integer}}
\item{Non-negative float = \code{numeric}}
\item{Currency amount = \code{numeric}}
\item{Float = \code{numeric}}
\item{Positive float = \code{numeric}}
\item{Latitude = \code{numeric}}
\item{Longitude = \code{numeric}}
}
}
\keyword{data}
