% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfsio.R
\docType{package}
\name{gtfsio}
\alias{gtfsio}
\alias{_PACKAGE}
\alias{gtfsio-package}
\title{gtfsio: Read and Write General Transit Feed Specification (GTFS) Data}
\description{
Fast and flexible tools to read and write General
    Transit Feed Specification (GTFS) files.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-transit/gtfsio}
  \item Report bugs at \url{https://github.com/r-transit/gtfsio/issues}
}

}
\author{
\strong{Maintainer}: Daniel Herszenhut \email{dhersz@gmail.com} (\href{https://orcid.org/0000-0001-8066-1105}{ORCID})

}
