% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfsio_error.R
\name{gtfsio_error}
\alias{gtfsio_error}
\title{gtfsio's custom error condition constructor}
\usage{
gtfsio_error(message, subclass = character(0), call = sys.call(-1))
}
\arguments{
\item{message}{The message to inform about the error.}

\item{subclass}{The subclass of the error.}

\item{call}{A call to associate the error with.}
}
\description{
gtfsio's custom error condition constructor
}
\seealso{
Other error constructors: 
\code{\link{parent_function_error}()}
}
\concept{error constructors}
\keyword{internal}
