% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_gtfs.R
\name{import_gtfs}
\alias{import_gtfs}
\title{Import GTFS transit feeds}
\usage{
import_gtfs(path, files = NULL, fields = NULL, quiet = TRUE)
}
\arguments{
\item{path}{A string. The path to a GTFS \code{.zip} file.}

\item{files}{A character vector. The text files to be read from the GTFS,
without the \code{.txt} extension. If \code{NULL} (the default), all
existing text files are read.}

\item{fields}{A named list. The fields to be read from each text file, in the
format \code{list(file1 = c("field1", "field2"))}. If \code{NULL} (the
default), all fields from the files specified in \code{files} are read. If
a file is specified in \code{files} but not in \code{fields}, all fields
from that file will be read (i.e. you may specify in \code{fields} only
files whose fields you want to subset).}

\item{quiet}{A logical. Whether to hide log messages and progress bars
(defaults to \code{TRUE}).}
}
\value{
A GTFS object: a named \code{list} of \code{data.table}s, in which
  each table represents a GTFS text file.
}
\description{
Imports GTFS transit feeds from either a local \code{.zip} file or an URL.
Columns are parsed according to the standards for reading and writing GTFS
feeds specified in \code{\link{get_gtfs_standards}}.
}
\seealso{
\code{\link{get_gtfs_standards}}
}
\concept{io}
