% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtfs.R
\name{new_gtfs}
\alias{new_gtfs}
\title{GTFS object constructor}
\usage{
new_gtfs(x, names, subclass = character(), ...)
}
\arguments{
\item{x}{A list. Each element must represent a distinct GTFS text file.}

\item{names}{A character vector. The name of each \code{gtfs} object element
(i.e. the name of the GTFS text files that each dataframe represents).}

\item{subclass}{A character vector. Subclasses to be assigned to the
\code{gtfs} object.}

\item{...}{Name-value pairs. Additional attributes.}
}
\value{
A GTFS object: a named list of dataframes, each one corresponding to
  a distinct GTFS text file.
}
\description{
Creates a GTFS object. Mostly useful for package authors who may want to
either create \code{gtfs} objects in their packages or create subclasses of
the main \code{gtfs} class. The usage of this function assumes some knowledge
on \code{gtfs} objects, thus inputs are not extensively checked. See
\code{gtfs} (TODO: remember to add link to gtfs()) for a user-friendlier and more thoroughly checked
constructor.
}
\examples{
gtfs_path <- system.file("extdata/ggl_gtfs.zip", package = "gtfsio")

tmpdir <- file.path(tempdir(), "new_gtfs_example")
zip::unzip(gtfs_path, exdir = tmpdir)

agency <- data.table::fread(file.path(tmpdir, "agency.txt"))
stops <- data.table::fread(file.path(tmpdir, "stops.txt"))
routes <- data.table::fread(file.path(tmpdir, "routes.txt"))
trips <- data.table::fread(file.path(tmpdir, "trips.txt"))
stop_times <- data.table::fread(file.path(tmpdir, "stop_times.txt"))
calendar <- data.table::fread(file.path(tmpdir, "calendar.txt"))

flist <- list(agency, stops, routes, trips, stop_times, calendar)
fname <- c("agency", "stops", "routes", "trips", "stop_times", "calendar")

gtfs <- new_gtfs(flist, fname)

class(gtfs)
names(gtfs)

}
\seealso{
\code{gtfs} (TODO: remember to add link to gtfs())
}
